Update MtLebo Spider
Modify some xpath addressing to make them more flexible and less likely to be disrupted when the formatting of the page changes. This formatting looks for keywords and expects certain elements to be positioned relative to those keywords.

Change nomenclature of constants to ALL_CAPS pattern.

Rename file/directory names.

Remove commented-out code not in use.