Fixing an issue with launching scrapy shell
This is a work-around to a None-reference that occurs in the base class
from the cityscrapers core module. The exception is raised from a factory
method that doesn't check for a null spider (which seems to occur when
you launch from scrapy shell). This fix overrides that factory method and
adds a None check.